Move CodecMode to amqp-message, reduce public API surface
CodecMode belongs on the message (the thing that's generic over it),
not the publisher. Stop exporting CodecMode, PublishBody, and
Serializable from index.ts since users never need to import them
directly; they're inferred from connect() and method signatures.
