+++ src/amqp-exchange.ts
@@ -1,7 +1,8 @@
 import type { AMQPProperties } from "./amqp-properties.js"
 import type { AMQPSession } from "./amqp-session.js"
 import { publishConfirmed, publishNoConfirm } from "./amqp-publisher.js"
-import type { PublishBody, Serializable, CodecMode } from "./amqp-publisher.js"
+import type { PublishBody, Serializable } from "./amqp-publisher.js"
+import type { CodecMode } from "./amqp-message.js"
 
 /** Options for {@link AMQPExchange#publish}. */
 export type ExchangePublishOptions = AMQPProperties & {
