+++ src/amqp-publisher.ts
@@ -1,5 +1,6 @@
 import type { AMQPProperties } from "./amqp-properties.js"
 import type { AMQPSession } from "./amqp-session.js"
+import type { CodecMode } from "./amqp-message.js"
 
 export type Body = string | Uint8Array | ArrayBuffer | Buffer | null
 
@@ -13,8 +14,6 @@ export function isBody(data: unknown): data is Body {
 }
 export type Serializable = Body | number | boolean | Record<string, unknown> | unknown[]
 
-/** Controls whether publish methods accept rich types or only raw bytes. */
-export type CodecMode = "plain" | "codec"
 export type PublishBody<C extends CodecMode> = C extends "codec" ? Serializable : Body
 
 /** Publish with broker confirmation. */
