+++ src/amqp-session.ts
@@ -3,7 +3,8 @@ import type { AMQPChannel, ExchangeParams, ExchangeType, QueueParams } from "./a
 import type { AMQPCodecRegistry } from "./amqp-codec-registry.js"
 import type { AMQPProperties } from "./amqp-properties.js"
 import { isBody } from "./amqp-publisher.js"
-import type { Body, Serializable, PublishBody, CodecMode } from "./amqp-publisher.js"
+import type { Body, Serializable, PublishBody } from "./amqp-publisher.js"
+import type { CodecMode } from "./amqp-message.js"
 import { AMQPQueue } from "./amqp-queue.js"
 import type { AMQPTlsOptions } from "./amqp-tls-options.js"
 import type { Logger } from "./types.js"
